I've resolved the CI failures by consolidating and removing redundant slider tests.
- I merged the `OptionSingleSlider` test logic into `OptionSliderControl.test.tsx`.
- I removed the obsolete `OptionSingleSlider.tsx` source and test files to ensure consistency.
- I corrected a minor `aria-label` rendering issue in `OptionSliderControl`.
- I have verified these changes using local test suites and visual checks.
