+++ src/components/ui/slider.tsx
@@ -25,6 +25,7 @@ function Slider({
 			min={min}
 			max={max}
 			thumbAlignment="edge"
+			aria-label={props["aria-label"]}
 			{...props}
 		>
 			<SliderPrimitive.Control className="relative flex w-full touch-none items-center select-none data-disabled:opacity-50 data-vertical:h-full data-vertical:min-h-40 data-vertical:w-auto data-vertical:flex-col">
