+++ tests/OptionSliderControl.test.tsx
@@ -5,6 +5,7 @@ import { OptionSliderControl } from "@/components/blocks/OptionSliderControl";
 describe("OptionSliderControl", () => {
 	const mockValues = [10, 20, 30, 40, 50];
 	const mockFormat = "{0}s";
+	const mockLabel = "Test Label";
 
 	it("renders slider and input with correct initial value", async () => {
 		await act(async () => {
@@ -28,6 +29,22 @@ describe("OptionSliderControl", () => {
 		expect(input).toHaveValue("20");
 
 		expect(screen.getByText("s")).toBeInTheDocument();
+	});
+
+	it("renders label when provided", async () => {
+		await act(async () => {
+			render(
+				<OptionSliderControl
+					label={mockLabel}
+					selection={1}
+					values={mockValues}
+					format={mockFormat}
+					onChange={() => {}}
+				/>,
+			);
+		});
+
+		expect(screen.getByText(mockLabel)).toBeInTheDocument();
 	});
 
 	it("calls onChange when slider moves", async () => {
@@ -70,7 +87,7 @@ describe("OptionSliderControl", () => {
 		await act(async () => {
 			fireEvent.change(input, { target: { value: "24" } });
 		});
-		// Should not be called yet (unified behavior: blur/enter)
+		// Should not be called yet
 		expect(onChange).not.toHaveBeenCalled();
 
 		await act(async () => {
@@ -91,5 +108,61 @@ describe("OptionSliderControl", () => {
 			fireEvent.blur(input);
 		});
 		expect(onChange).toHaveBeenCalledWith(4);
+	});
+
+	it("allows clearing input and then entering a value", async () => {
+		const onChange = vi.fn();
+		await act(async () => {
+			render(
+				<OptionSliderControl
+					selection={1}
+					values={mockValues}
+					format={mockFormat}
+					onChange={onChange}
+				/>,
+			);
+		});
+
+		const input = screen.getByRole("textbox");
+
+		await act(async () => {
+			fireEvent.change(input, { target: { value: "" } });
+		});
+		expect(input).toHaveValue("");
+		expect(onChange).not.toHaveBeenCalled();
+
+		await act(async () => {
+			fireEvent.change(input, { target: { value: "35" } });
+			fireEvent.blur(input);
+		});
+		// 35 is between 30 (index 2) and 40 (index 3).
+		// values = [10, 20, 30, 40, 50]
+		expect(onChange).toHaveBeenCalledWith(2);
+	});
+
+	it("snaps input value back even if selection index doesn't change", async () => {
+		const onChange = vi.fn();
+		await act(async () => {
+			render(
+				<OptionSliderControl
+					selection={1} // Value is 20
+					values={mockValues}
+					format={mockFormat}
+					onChange={onChange}
+				/>,
+			);
+		});
+
+		const input = screen.getByRole("textbox");
+
+		// 21 is closest to 20 (index 1)
+		await act(async () => {
+			fireEvent.change(input, { target: { value: "21" } });
+			fireEvent.blur(input);
+		});
+
+		// Selection index 1 didn't change, but input should snap back to "20"
+		expect(onChange).toHaveBeenCalledWith(1);
+		expect(input).toHaveValue("20");
 	});
 });
